requirements.txt
Django>=5.2.3
psycopg2>=2.9.9
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1 